whip/src/adapter/requirements.txt
evernote==1.25.0
Flask==0.10.1
httplib2==0.9
itsdangerous==0.24
Jinja2==2.8
lru-dict==1.1.6
lxml==3.3.5
Mako==0.9.0
MarkupSafe==0.18
nose==1.3.4
oauth2==1.5.211
ply==3.4
python-chess==0.7.0
pytz==2014.4
PyYAML==3.11
spyne==2.10.10
suds-jurko==0.6
thrift==0.9.1
Werkzeug==0.10.4

